Aggregate tags differently for each reader
