+++ src/Reader/AbstractReader.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Azura\MetadataManager\Reader;
 
-use Azura\MetadataManager\Utilities\Arrays;
 use voku\helper\UTF8;
 
 abstract class AbstractReader implements ReaderInterface
@@ -14,35 +13,6 @@ abstract class AbstractReader implements ReaderInterface
         string $jsonOutput,
         ?string $artOutput
     ): void;
-
-    protected static function aggregateMetaTags(array $toProcess): array
-    {
-        $metaTags = [];
-
-        foreach ($toProcess as $tagSet) {
-            if (empty($tagSet)) {
-                continue;
-            }
-
-            foreach ($tagSet as $tagName => $tagContents) {
-                if (!empty($tagContents[0]) && !isset($metaTags[$tagName])) {
-                    $tagValue = $tagContents[0];
-                    if (is_array($tagValue)) {
-                        // Skip pictures
-                        if (isset($tagValue['data'])) {
-                            continue;
-                        }
-                        $flatValue = Arrays::flattenArray($tagValue);
-                        $tagValue = implode(', ', $flatValue);
-                    }
-
-                    $metaTags[(string)$tagName] = self::cleanUpString((string)$tagValue);
-                }
-            }
-        }
-
-        return $metaTags;
-    }
 
     protected static function cleanUpString(?string $original): string
     {
