+++ src/Reader/FfmpegReader.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Azura\MetadataManager\Reader;
 
 use Azura\MetadataManager\Metadata;
+use Azura\MetadataManager\Utilities\Arrays;
 use Azura\MetadataManager\Utilities\Time;
 use FFMpeg\FFMpeg;
 use FFMpeg\FFProbe;
@@ -40,6 +41,8 @@ class FfmpegReader extends AbstractReader
 
         $metaTags = self::aggregateMetaTags($toProcess);
 
+        file_put_contents(__DIR__ . '/test1.json', json_encode($metaTags, JSON_PRETTY_PRINT));
+
         $metadata->setTags($metaTags);
         $metadata->setMimeType(mime_content_type($path) ?: '');
 
@@ -69,5 +72,34 @@ class FfmpegReader extends AbstractReader
                 break;
             }
         }
+    }
+
+    protected static function aggregateMetaTags(array $toProcess): array
+    {
+        $metaTags = [];
+
+        foreach ($toProcess as $tagSet) {
+            if (empty($tagSet)) {
+                continue;
+            }
+
+            foreach ($tagSet as $tagName => $tagContents) {
+                if (!empty($tagContents) && !isset($metaTags[$tagName])) {
+                    $tagValue = $tagContents;
+                    if (is_array($tagValue)) {
+                        // Skip pictures
+                        if (isset($tagValue['data'])) {
+                            continue;
+                        }
+                        $flatValue = Arrays::flattenArray($tagValue);
+                        $tagValue = implode(', ', $flatValue);
+                    }
+
+                    $metaTags[(string)$tagName] = self::cleanUpString((string)$tagValue);
+                }
+            }
+        }
+
+        return $metaTags;
     }
 }
