+++ src/Reader/GetId3Reader.php
@@ -6,6 +6,7 @@ namespace Azura\MetadataManager\Reader;
 
 use Azura\MetadataManager\Exception\ReadException;
 use Azura\MetadataManager\Metadata;
+use Azura\MetadataManager\Utilities\Arrays;
 use Azura\MetadataManager\Utilities\Time;
 use JamesHeinrich\GetID3\GetID3;
 
@@ -79,5 +80,34 @@ class GetId3Reader extends AbstractReader
                 );
             }
         }
+    }
+
+    protected static function aggregateMetaTags(array $toProcess): array
+    {
+        $metaTags = [];
+
+        foreach ($toProcess as $tagSet) {
+            if (empty($tagSet)) {
+                continue;
+            }
+
+            foreach ($tagSet as $tagName => $tagContents) {
+                if (!empty($tagContents[0]) && !isset($metaTags[$tagName])) {
+                    $tagValue = $tagContents[0];
+                    if (is_array($tagValue)) {
+                        // Skip pictures
+                        if (isset($tagValue['data'])) {
+                            continue;
+                        }
+                        $flatValue = Arrays::flattenArray($tagValue);
+                        $tagValue = implode(', ', $flatValue);
+                    }
+
+                    $metaTags[(string)$tagName] = self::cleanUpString((string)$tagValue);
+                }
+            }
+        }
+
+        return $metaTags;
     }
 }
